+++ internal/utils/packetID.go
@@ -4,7 +4,7 @@ import (
 	"sync"
 )
 
-func ReadAndIncrementPacketID(packetID *byte, packetMutex *sync.Mutex, increment bool) byte {
+func ReadAndIncrementPacketID(packetID *uint16, packetMutex *sync.Mutex, increment bool) uint16 {
 	packetMutex.Lock()
 	defer packetMutex.Unlock()
 	id := *packetID
